examples/docker/api_requirements.txt
# API Server Requirements
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.26.0
pyyaml>=6.0.0
python-multipart>=0.0.6
